Bug fix for affine gap cost problem with profile alignment
Sometimes the option with more gap openings were chosen because we were not able to detect all gap open events. The solution was to simplify the code and do not make the (incomplete) attempt to assess already present gap characters with affine gap costs. Instead always gap extension is now chosen for those gap characters.